Refactor UI into modules and add tests
